[dbsp] Rearrange `trace::spine_async` a little.
`IndexSet` will be used outside of `trace::spine_async::list_merger` in an
upcoming commit, so move it to a module of its own.
